# contracts/order-service-openapi.yaml
openapi: 3.1.0
info:
  title: Order Service API
  summary: Create and retrieve customer orders.
  description: |
    HTTP contract for the first version of the order service.

    Orders are created in PENDING status. The service currently stores the
    order and its lines; event publication is introduced in a later phase.
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: Orders
    description: Order creation and retrieval
paths:
  /api/v1/orders:
    post:
      tags:
        - Orders
      operationId: createOrder
      summary: Create an order
      description: |
        Creates a PENDING order. The Idempotency-Key must be reused when
        retrying the same request. Reusing it with different content returns
        a conflict.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            examples:
              basicOrder:
                summary: Basic order with one line
                value:
                  customerId: ad2f3742-a4a6-4b9c-aed1-45884cc7bd21
                  lines:
                    - productId: 55db5caf-b0da-49ab-b346-b7581a91ee76
                      quantity: 2
                      unitPrice: 19.99
      responses:
        '201':
          description: Order created.
          headers:
            Location:
              $ref: '#/components/headers/Location'
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/BusinessValidationError'

  /api/v1/orders/{orderId}:
    get:
      tags:
        - Orders
      operationId: getOrder
      summary: Get an order
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Order found.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/OrderNotFound'

components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key used to safely retry an order creation request.
      schema:
        type: string
        minLength: 1
        maxLength: 255
      example: demo-order-1
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Optional request identifier. The service generates one when absent.
      schema:
        type: string
        minLength: 1
        maxLength: 100
      example: postman-order-test
    OrderId:
      name: orderId
      in: path
      required: true
      description: Order identifier.
      schema:
        type: string
        format: uuid

  headers:
    Location:
      description: URL of the newly created order.
      schema:
        type: string
        format: uri-reference
      example: /api/v1/orders/2d931510-d99f-494a-8c67-87feb05e1594
    CorrelationId:
      description: Correlation identifier assigned to the request.
      schema:
        type: string
      example: postman-order-test

  responses:
    BadRequest:
      description: The request is syntactically invalid or cannot be parsed.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://distributed-orders.example/problems/400
            title: Request validation failed
            status: 400
            detail: One or more request fields are invalid
            instance: /api/v1/orders
            correlationId: postman-order-test
            fieldErrors:
              - field: lines
                message: must not be empty
    OrderNotFound:
      description: The requested order does not exist.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    IdempotencyConflict:
      description: The idempotency key was used with different content.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            type: https://distributed-orders.example/problems/409
            title: Idempotency conflict
            status: 409
            detail: The idempotency key was used with different content
            instance: /api/v1/orders
            correlationId: postman-order-test
    BusinessValidationError:
      description: The request violates a business rule.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

  schemas:
    CreateOrderRequest:
      type: object
      required:
        - customerId
        - lines
      properties:
        customerId:
          type: string
          format: uuid
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/CreateOrderLineRequest'
    CreateOrderLineRequest:
      type: object
      required:
        - productId
        - quantity
        - unitPrice
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          minimum: 1
        unitPrice:
          type: number
          format: double
          minimum: 0.01
          example: 19.99
    OrderResponse:
      type: object
      required:
        - orderId
        - customerId
        - status
        - lines
        - total
        - currency
        - createdAt
        - updatedAt
      properties:
        orderId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - PENDING
            - STOCK_RESERVED
            - PAYMENT_PENDING
            - PAYMENT_AUTHORIZED
            - CONFIRMED
            - CANCELLING
            - CANCELLED
        lines:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineResponse'
        total:
          type: number
          format: double
          minimum: 0
          example: 39.98
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          example: EUR
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OrderLineResponse:
      type: object
      required:
        - productId
        - quantity
        - unitPrice
        - subtotal
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          minimum: 1
        unitPrice:
          type: number
          format: double
          minimum: 0
        subtotal:
          type: number
          format: double
          minimum: 0
    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - instance
        - correlationId
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        correlationId:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
      additionalProperties: true
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string