# services/order-service/src/main/java/com/robert/portfolio/orders/infrastructure/web/OrderController.java
package com.robert.portfolio.orders.infrastructure.web;

import com.robert.portfolio.orders.application.port.in.CreateOrderUseCase;
import com.robert.portfolio.orders.application.port.in.GetOrderUseCase;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.net.URI;
import java.util.UUID;

@RestController
@RequestMapping("/api/v1/orders")
public class OrderController {

    private final CreateOrderUseCase createOrderUseCase;
    private final GetOrderUseCase getOrderUseCase;
    private final OrderWebMapper mapper;

    public OrderController(CreateOrderUseCase createOrderUseCase,
                           GetOrderUseCase getOrderUseCase,
                           OrderWebMapper mapper) {
        this.createOrderUseCase = createOrderUseCase;
        this.getOrderUseCase = getOrderUseCase;
        this.mapper = mapper;
    }

    @PostMapping
    public ResponseEntity<OrderResponse> createOrder(
            @RequestHeader("Idempotency-Key") String idempotencyKey,
            @Valid @RequestBody CreateOrderRequest request) {
        var command = mapper.toCommand(request, idempotencyKey);
        OrderResponse response = mapper.toResponse(createOrderUseCase.create(command));
        return ResponseEntity.created(URI.create("/api/v1/orders/" + response.orderId())).body(response);
    }

    @GetMapping("/{orderId}")
    public OrderResponse getOrder(@PathVariable UUID orderId) {
        return mapper.toResponse(getOrderUseCase.get(orderId));
    }
}